Move noise model from t.geometry to t.io
