+++ cpp/open3d/t/geometry/Image.h
@@ -357,38 +357,6 @@ protected:
     core::Tensor data_;
 };
 
-/// Simulate depth image noise from a given noise distortion model.
-class DepthNoiseSimulator {
-public:
-    /// \brief Constructor.
-    /// \param noise_model Path to the noise model file. See
-    /// http://redwood-data.org/indoor/dataset.html for the format. Or, you may
-    /// use one of our example datasets, e.g., RedwoodIndoorLivingRoom1.
-    DepthNoiseSimulator(const std::string &noise_model_path);
-
-    /// \brief Apply noise model to a depth image.
-    ///
-    /// \param im_src Source depth image, must be with dtype UInt16 or
-    /// Float32, channels==1.
-    /// \param depth_scale Scale factor to the depth image. As a sanity check,
-    /// if the dtype is Float32, the depth_scale must be 1.0. If the dtype is
-    /// is UInt16, the depth_scale is typically larger than 1.0, e.g. it can be
-    /// 1000.0.
-    /// \return Noisy depth image with the same shape and dtype as \p im_src.
-    Image Simulate(const Image &im_src, float depth_scale = 1000.0);
-
-    /// \brief Return the noise model.
-    core::Tensor GetNoiseModel() const { return model_; }
-
-    /// \brief Enable deterministic debug mode. All normally distributed noise
-    /// will be replaced by 0.
-    void EnableDeterministicDebugMode() { deterministic_debug_mode_ = true; }
-
-private:
-    core::Tensor model_;  // ndims==3
-    bool deterministic_debug_mode_ = false;
-};
-
 }  // namespace geometry
 }  // namespace t
 }  // namespace open3d
