+++ cpp/pybind/t/io/class_io.cpp
@@ -184,6 +184,65 @@ void pybind_class_io(py::module &m_io) {
             "write_triangle_uvs"_a = true, "print_progress"_a = false);
     docstring::FunctionDocInject(m_io, "write_triangle_mesh",
                                  map_shared_argument_docstrings);
+
+    // DepthNoiseSimulator
+    py::class_<DepthNoiseSimulator> depth_noise_simulator(
+            m_io, "DepthNoiseSimulator",
+            R"(Simulate depth image noise from a given noise distortion model. The distortion model is based on *Teichman et. al. "Unsupervised intrinsic calibration of depth sensors via SLAM" RSS 2009*. Also see <http://redwood-data.org/indoor/dataset.html>__
+
+Example::
+
+    import open3d as o3d
+
+    # Redwood Indoor LivingRoom1 (Augmented ICL-NUIM)
+    # http://redwood-data.org/indoor/
+    data = o3d.data.RedwoodIndoorLivingRoom1()
+    noise_model_path = data.noise_model_path
+    im_src_path = data.depth_paths[0]
+    depth_scale = 1000.0
+
+    # Read clean depth image (uint16)
+    im_src = o3d.t.io.read_image(im_src_path)
+
+    # Run noise model simulation
+    simulator = o3d.t.io.DepthNoiseSimulator(noise_model_path)
+    im_dst = simulator.simulate(im_src, depth_scale=depth_scale)
+
+    # Save noisy depth image (uint16)
+    o3d.t.io.write_image("noisy_depth.png", im_dst)
+            )");
+    depth_noise_simulator.def(py::init<const std::string &>(),
+                              "noise_model_path"_a);
+    depth_noise_simulator.def("simulate", &DepthNoiseSimulator::Simulate,
+                              "im_src"_a, "depth_scale"_a = 1000.0f,
+                              "Apply noise model to a depth image.");
+    depth_noise_simulator.def(
+            "enable_deterministic_debug_mode",
+            &DepthNoiseSimulator::EnableDeterministicDebugMode,
+            "Enable deterministic debug mode. All normally distributed noise "
+            "will be replaced by 0.");
+    depth_noise_simulator.def_property_readonly(
+            "noise_model", &DepthNoiseSimulator::GetNoiseModel,
+            "The noise model tensor.");
+    docstring::ClassMethodDocInject(
+            m_io, "DepthNoiseSimulator", "__init__",
+            {{"noise_model_path",
+              "Path to the noise model file. See "
+              "http://redwood-data.org/indoor/dataset.html for the format. Or, "
+              "you may use one of our example datasets, e.g., "
+              "RedwoodIndoorLivingRoom1."}});
+    docstring::ClassMethodDocInject(
+            m_io, "DepthNoiseSimulator", "simulate",
+            {{"im_src",
+              "Source depth image, must be with dtype UInt16 or Float32, "
+              "channels==1."},
+             {"depth_scale",
+              "Scale factor to the depth image. As a sanity check, if the "
+              "dtype is Float32, the depth_scale must be 1.0. If the dtype is "
+              "is UInt16, the depth_scale is typically larger than 1.0, e.g. "
+              "it can be 1000.0."}});
+    docstring::ClassMethodDocInject(m_io, "DepthNoiseSimulator",
+                                    "enable_deterministic_debug_mode");
 }
 
 }  // namespace io
